news_publisher/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-logging
packaging
requests
